requirements.txt
langchain==0.1.11
gradio==5.23.2
transformers==4.38.2
bs4==0.0.2
requests==2.31.0
torch==2.2.1
Pillow==10.2.0